# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Hotspot captive portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Package {
  id: number;
  name: string;
  /** Duration in hours */
  durationHours: number;
  /** Data limit in MB (0 = unlimited) */
  dataLimitMb: number;
  /** Price in KES */
  price: number;
  /** @nullable */
  description?: string | null;
  isActive?: boolean;
  createdAt?: string;
}

export interface PackageInput {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  durationHours: number;
  /** @minimum 0 */
  dataLimitMb: number;
  /** @minimum 1 */
  price: number;
  description?: string;
  isActive?: boolean;
}

export type SessionStatus = typeof SessionStatus[keyof typeof SessionStatus];


export const SessionStatus = {
  pending: 'pending',
  paid: 'paid',
  expired: 'expired',
} as const;

export interface Session {
  id: number;
  phone: string;
  packageId: number;
  /** @nullable */
  packageName?: string | null;
  /** @nullable */
  packagePrice?: number | null;
  status: SessionStatus;
  /** @nullable */
  paymentReference?: string | null;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  expiresAt?: string | null;
  createdAt: string;
}

export interface SessionInput {
  /** @minLength 9 */
  phone: string;
  packageId: number;
  ipAddress?: string;
  /** If true, session is immediately marked as paid (admin bypass) */
  adminCreated?: boolean;
  /**
     * Number of device slots to create (admin only — creates numDevices paid sessions)
     * @minimum 1
     * @maximum 20
     */
  numDevices?: number;
}

export interface PhoneLoginInput {
  /** @minLength 9 */
  phone: string;
  ipAddress: string;
}

export interface SessionStats {
  total: number;
  paid: number;
  pending: number;
  expired: number;
  revenue: number;
}

export interface PaidIpEntry {
  sessionId: number;
  ipAddress: string;
  phone?: string;
  expiresAt: string;
}

export interface PaymentInitiate {
  sessionId: number;
  phone: string;
  amount: number;
}

export interface PaymentResponse {
  success: boolean;
  message: string;
  /** @nullable */
  reference?: string | null;
}

export interface PaymentCallbackInput {
  status?: string;
  reference?: string;
  phone?: string;
  amount?: number;
}

export type PaymentStatusResultStatus = typeof PaymentStatusResultStatus[keyof typeof PaymentStatusResultStatus];


export const PaymentStatusResultStatus = {
  pending: 'pending',
  paid: 'paid',
  failed: 'failed',
} as const;

export interface PaymentStatusResult {
  status: PaymentStatusResultStatus;
  /** @nullable */
  sessionId?: number | null;
}

export type VisitorDeviceType = typeof VisitorDeviceType[keyof typeof VisitorDeviceType];


export const VisitorDeviceType = {
  mobile: 'mobile',
  tablet: 'tablet',
  desktop: 'desktop',
} as const;

export interface Visitor {
  ipAddress: string;
  userAgent: string;
  deviceType: VisitorDeviceType;
  os: string;
  browser: string;
  firstSeenAt: string;
  lastSeenAt: string;
}

export interface VisitorPingInput {
  ipAddress: string;
  userAgent?: string;
}

export type PingVisitor200 = {
  ok?: boolean;
};