# lib/api-zod/src/generated/types/package.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Hotspot captive portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Package {
  id: number;
  name: string;
  /** Duration in hours */
  durationHours: number;
  /** Data limit in MB (0 = unlimited) */
  dataLimitMb: number;
  /** Price in KES */
  price: number;
  /** @nullable */
  description?: string | null;
  isActive?: boolean;
  createdAt?: string;
}