# lib/api-zod/src/generated/types/packageInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Hotspot captive portal API
 * OpenAPI spec version: 0.1.0
 */

export interface PackageInput {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  durationHours: number;
  /** @minimum 0 */
  dataLimitMb: number;
  /** @minimum 1 */
  price: number;
  description?: string;
  isActive?: boolean;
}