# lib/api-zod/src/generated/types/paymentCallbackInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Hotspot captive portal API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentCallbackInput {
  status?: string;
  reference?: string;
  phone?: string;
  amount?: number;
}