# lib/api-zod/src/generated/types/paymentStatusResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Hotspot captive portal API
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentStatusResultStatus } from './paymentStatusResultStatus';

export interface PaymentStatusResult {
  status: PaymentStatusResultStatus;
  /** @nullable */
  sessionId?: number | null;
}