# lib/api-zod/src/generated/types/sessionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Hotspot captive portal API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionInput {
  /** @minLength 9 */
  phone: string;
  packageId: number;
  ipAddress?: string;
  /** If true, session is immediately marked as paid (admin bypass) */
  adminCreated?: boolean;
  /**
     * Number of device slots to create (admin only — creates numDevices paid sessions)
     * @minimum 1
     * @maximum 20
     */
  numDevices?: number;
}