# lib/api-zod/src/generated/types/sessionStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Hotspot captive portal API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionStats {
  total: number;
  paid: number;
  pending: number;
  expired: number;
  revenue: number;
}