# lib/api-zod/src/generated/types/visitorPingInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Hotspot captive portal API
 * OpenAPI spec version: 0.1.0
 */

export interface VisitorPingInput {
  ipAddress: string;
  userAgent?: string;
}